feat(request): add hubRequest() for making requests to the Hub API
AFFECTS PACKAGES:
@esri/hub-common
@esri/hub-content
